api: persist db-backed training pipeline

+++ migrations/versions/202503041200_initial_schema.py
@@ -120,7 +120,7 @@ def upgrade() -> None:
 
     op.create_table(
         "dataset_images",
-        sa.Column("id", sa.BigInteger(), primary_key=True, nullable=False),
+        sa.Column("id", sa.Integer(), primary_key=True, nullable=False),
         sa.Column(
             "dataset_id",
             sa.Integer(),
